9p: Reorganize the mount parsing code.
To compare or test mount parameters for equivalence, it is necessary
to have all the mount parameters apriori so that they can be compared
with the mount options of existing superblocks. This is important
specially when a cache is to be shared between similar mounts so that
we do not end up having two copies of the cache.